Move stuff from common/src/lib.rs to common/src/rv_trace.rs
